Working around bundler & rpms.
(1) `bundler install --local` can't ignore groups. In order to get
a complete gemset (a.k.a. a Gemfile.lock) for production, I need a separate
Gemfile for production than development.

(2) We don't need to use bundler in the app; the actual gem list is short
enough to handle with `require` statements.

(3) We won't *need* bundler & the Gemfile in production at all; the value in
having it, is in generating a finite list of required gems.