Add orm_exporttopostgres to bin/
Port the orm_exporttopostgres script to work with the new pint_server
implementation.

To use it you should clone the pint-data repo, that contains the XML
data to be import, beside the public-cloud-info-service; the script
will automatically look for the pint-data repo relative to it's own
location in the public-cloud-info-service/bin directory.

Run it by specifying the necessary command line arguments, e.g.

    % bin/orm_exporttopostgres --host ${POSTGRES_HOST} \
        --password ${POSTGRES_PASSWORD}
        --user ${POSTGRES_USER}

The script will create the database tables if not already created,
and will then import the XML content from the pint-data repo, either
as a full import if no data exists, or incrementally importing new
changes made since the last import.

For incremental updates, when updating the versions table, if the
existing version entry value is less than the detected commit date
stamp for the updates, then we will use the new commit date stamp,
otherwise we will add 0.01 to the existing date stamp value. This
should allow us to support up to 100 incremental merges for the
same date without issues, and will still work if we ever have more
than that number of incremental updates, though the relationship
between the version value and the associated commit data may be
lost.

To test the incremental update support I used the following shell
command when cd'd into a pint data repo clone, with this repo
cloned beside it:

    % for i in $(seq 90 -1 0); do git co origin/master~${i} && \
      ../public-cloud-info-service/bin/orm_exporttopostgres \
          --host ${POSTGRES_HOST} \
          --password ${POSTGRES_PASSWORD} \
          --user ${POSTGRES_USER}; done

Include lmxl (v4.6.3) in the requirements.txt to support running
the orm_exporttopostgres script.