add schema versioning and migration framework
The schema versioning and migration is facilitated by sqlalchemy-migrate[1].

Also add a CLI (implemented using python-click) to simpify schema and data
upgrade. With the CLI, we no longer need orm_exporttogres script and the code
have be incorporated into the CLI as well. To perform schema and data upgrade,
we simply need to invoke the pint_db_migrate.sh CLI. For example:

./bin/pint_db_migrate.sh -h db_host -U db_user -W db_password upgrade

TODO: We will need to remove the orm_exporttopostgres script once we are OK
with this PR.

1. https://sqlalchemy-migrate.readthedocs.io/en/latest/api.html