Implement init_db() enhancements in pint_server
Add support to the pint_server init_db() function to take optional
dbconfig, outputfile and echo parameters.

The dbconfig paramater allows a dictionary to be specified that
holds the required config settings needed to contruct the connection
URL string. If a dbconfig is specified then it's settings will be
used in preference to those specified in the environment. Otherwise
the DATABASE_URI env var setting will be used, or if that is not
specified then the individual POSTGRES_... env var settings will be
used.

If the outputfile parameter is specified it will be used to create
a DB specific logging service to log output from SQLAlchemy DB ops.

The echo parameter will be passed as the echo parameter when calling
create_engine() allowing control over the associated engine logging
verbosity.

Updated pint_server/__init__.py to explicitly import the objects
that we want to expose from the pint_server module.

Updated PintBase to provide a tablename property.