Implement a check mode
Implemented a "--check-mode" option for the upgrade command. If specified,
pint_db_migrate will simply perform data sanity check (i.e. duplicates)
and skip the actual data import. If the option is not specified,
duplicates will be ignored, with a warning message posted on the screen,
but the data import will proceed.

This PR also account for any empty fields during data import.