Database Re-Structuring
Update primary key definitions
==============================

Now that we have identified the target primary key definitions we can
set about implementing them as a series of Alembic migrations.

As part of handling this we have to make the following types of changes
to a number of tables:
  * add a new sequence based column to act as the primary key
  * update nullable settings for some fields, either because
    the existing settings are wrong or because they should be
    nullable, but previously were being used in primary key
    definitions and as such couldn't be configured as nullable.
  * change the primary key definition, potentially to an added
    sequence field.

This patch adds a new migration step which implements primary key
changes.

Custom Data Validation
======================

This patch adds custom validation of the servers tables name field,
ensuring that the name field cannot be a null/empty value for an
update server.

XML Data Import Updates
=======================

Fixed incremental update to work with tables whose primary key has
been converted to an integer sequence; the search for existing rows
will look for an exact match for all data fields rather that just
the primary key fields.

Misc Changes
============

Removed defunct orm_exporttopostgres script.

Updated run_standalone.sh to support overriding env settings so as
to avoid the need to edit a committed file to test.

Added git ignorance of the top-level .vscode directory.

Version bumped to 2.0.4

Closes: #74