Fix incremental update for microsoftimages table
The recent schema updates to change the primary keys associated with
some fields, and associated incremental update changes to deal with
servers tables now using a sequence column for primary keys, broke
the incremental update support for the microsoftimages table.

To fix this we need to specify an appropriate set of columns that can
be used for existing entry checks in this table, so as to ensure we
detect when an existing entry is being updated, rather than an new
entry being added.

The new IDENTITY_OVERRIDES mapping has been added, currently with a
single entry for the microsoftimages table, through more can be added
if needed, to support this fix. It can be extended to support similar
overrides for other tables should the need arrise.

Closes: #93