Add unique constraint to MicrosoftImagesModel
Add a UniqueConstraint based upon the name and environment columns to
the MicrosoftImagesModel (microsoftimages) table definition.

Add a schema migration to upgrade the database to add the new unique
constraint to the microsoftimages.

Add a new class method to the PintBase class that returns the list of
defined unique constraints associated with a given table, or an empty
list if none are found.

Update the data_update.py IDENTITY_OVERRIDES mechanism to leverage the
new unique_constraints class method from PintBase when specifying an
override entry for the 'microsoftimages' table.

Bump the version to 2.0.5

Closes: #97