changeinfo URLs should end with a /
Add a validator that ensures that the changeinfo URLs end in a /.

Also add a data migration step which will make the same change to all
the existing changeinfo URLs in the database for all of the provider
images tables.

Update the data_update.py XML data extraction to automatically add a
trailing slash, if not present, to the changeinfo entries loaded from
the XML data, ensuring that future incremental data updates from XML
data don't incorrectly report data updates when processing older
entries that didn't end with '/'.

Bump version to 2.0.9.

Relates: #98