Exclude new primary key id fields from REST API request results
The recent changes to add sequence based id fields to the Amazon,
Google and Microsoft servers tables, and the Microsoft images table
result in extraneous id fields appearing in the REST API request
results for these entities.

We can leverage the attr exclusion mechanism that is available in
the get_formatted_dict() helper function to avoid this problem.

Closes: #101