Add validation of image date fields
The 3 date fields in an images table entry should conform to the
following ordering rule, if not Null:

    publishedon <= deprecatedon <= deletedon

The publishedon field cannot be Null, but the others may be Null if
the image hasn't transitioned to the associated state yet.

While no schema or data migration is involved we still add an Alembic
migration script which checks the existing DB content to ensure that
all images entries conform to this rule, reporting any problem entries
and failing with a ValueError exception if errors occur.

Bump pint_server version to 2.0.7.

Closes: #104