Ensure uniqueness for the server entries in the servers tables
A server is represented by a unique combination of IP addresses (and
optionally a region in the Microsoft case) so we want to ensure that
we can only add a given server once to the Amazon and Google servers
tables, and once per region to the Microsoft servers table.

In particular if we have a server with IPv4 and IPv6 addresses X and
Y respectively, we shouldn't be able to add more than one server for
for each of those addresses, so attempting to add a server with IPv4
address X, and another with IPv6 address Y, and a third server with
both addresses should fail.

To achieve this we can leverage partial unique indices to ensure that
a given address only occurs once, or once per region in the Microsoft
case.

Added a new 'adhoc_db' utility command in the bin directory that can
be used for performing adhoc DB operations, such as running a test
for this duplicate servers entries scenario.

Bump the version to 2.0.8.

Closes: #100