Handle publishedon not yet set in data validator
When validating the publishedon, deprecatedon, and deletedon fields
we need to wait until published on has been set before trying to do
any validations against it.

Since the same validation is performed when triggered for any of the
three fields, it needs to support being called for deprecatedon or
deletedon before publishedon without failing, while performing all
of the validation when called for publishedon.

Also ensure that the schema_upgrade.py and data_update.py scripts
exit with a failed exit status if a failure occurs.

Closes: #104 (again), #107