IP Addresses must be unique per region
Previously we defined new partial unique indices to ensure that IP
addresses were uniquely specified in the Amazon, Google and Microsoft
servers tables, though we allowed for them being unique per region in
the case of Microsoft.

However, we sometimes temporarily specify the update servers from an
existing region as the update servers for a newly created region until
the update infrastructure for the new region is in place.

To address this operational requirement we need to include the region
field in the partial unique indices for all servers tables, not just
the Microsoft servers table.

Update the adhoc_db duplicate_servers test to reflect the fact that
all servers tables, rather than just the Microsoft servers table,
should have the second region with the repeated IP address entry
added to them.

Cleanup some comments in the report_duplicates() helper method and
extend it to take as argument the index of the field to be counted,
defaulting to the first field, rather than being hardcoded to use
the first field.

Closes: #100 (again)