Fix XML results formatting issues
The deletiondate response should be an empty deletiondate value
for not found, or not yet deprecated images, so that an appopriately
formatted XML response can be generated.

The XML result for the v1/images/deletedby/{data} request was broken
and the result needed some tweaking to generate an XML result.

However, further work is still required to ensure correct formatting
for the result, which needs to be a two level data hierarchy with a
list of provider entries each containing an optionally empty list of
images.