Function testing fixes and enhancements
As part of implementing functional tests for the new deletedby and
deletion date requests I identified that a number of existing tests
were not actually doing anything; iterating over the response for a
requests.get() is like iterating over an empty list.

As part of fixing the issues with the existing tests I identified a
number of common code patterns related to obtaining lists of servers,
regions, and image states using the V1 PINT API, which I moved to
helper functions that the tests now call when needed.

Simplified a couple of long conditional checks to check for a value
being in a list.

Added additional positive and negative tests for the new deletedby
and deletion date requests.